Use anonymous auth and remove sign-in
