+++ __tests__/index.test.tsx
@@ -1,23 +1,63 @@
-import { render, screen } from '@testing-library/react';
+import { render, screen, act } from '@testing-library/react';
 
-jest.mock('@/lib/firebaseClient', () => ({ auth: {} }));
+document.createRange = () => ({
+  setStart: () => {},
+  setEnd: () => {},
+  commonAncestorContainer: document.createElement('div')
+} as any);
 
-jest.mock('firebase/auth', () => ({
-  signInWithEmailAndPassword: jest.fn(),
-  createUserWithEmailAndPassword: jest.fn()
+jest.mock('@/lib/firebaseClient', () => ({
+  auth: {},
+  db: {},
 }));
 
-jest.mock('next/router', () => ({
-  useRouter: () => ({ push: jest.fn() })
-}));
-
-const Home = require('@/pages/index').default;
-
-test('renders login form', () => {
-  render(<Home />);
-  expect(screen.getByText('Recipe Tracker')).toBeInTheDocument();
-  expect(screen.getByPlaceholderText('Email')).toBeInTheDocument();
-  expect(screen.getByPlaceholderText('Password')).toBeInTheDocument();
-  expect(screen.getByText('Log In')).toBeInTheDocument();
-  expect(screen.getByText('Sign Up')).toBeInTheDocument();
+jest.mock('@/lib/mealsStore', () => {
+  const mockMeal = {
+    id: '1',
+    mealName: 'Pizza',
+    date: {
+      toDate: () => new Date('2024-01-02'),
+      toMillis: () => new Date('2024-01-02').getTime(),
+    },
+    pending: false,
+  };
+  return {
+    getAllMeals: jest.fn().mockResolvedValue([mockMeal]),
+    saveMeal: jest.fn().mockResolvedValue(undefined),
+    getPendingMeals: jest.fn().mockResolvedValue([]),
+    markMealSynced: jest.fn().mockResolvedValue(undefined),
+  };
 });
+
+const mockOnSnapshot = jest.fn((_q, cb) => {
+  cb({
+    docs: [
+      { id: '1', data: () => ({ mealName: 'Pizza', date: { toDate: () => new Date('2024-01-02') } }) }
+    ]
+  });
+  return jest.fn();
+});
+
+jest.mock('firebase/firestore', () => ({
+  collection: jest.fn(),
+  addDoc: jest.fn(),
+  query: jest.fn(),
+  orderBy: jest.fn(),
+  onSnapshot: mockOnSnapshot,
+  Timestamp: {
+    fromDate: (d: Date) => ({ toDate: () => d, toMillis: () => d.getTime() }),
+    now: () => ({ toDate: () => new Date('2024-01-03'), toMillis: () => new Date('2024-01-03').getTime() })
+  }
+}));
+
+const Page = require('@/pages/index').default;
+
+test('renders meals list', async () => {
+  await act(async () => {
+    render(<Page />);
+  });
+  const dateString = new Date('2024-01-02').toLocaleDateString();
+  expect(screen.getByText('Meals')).toBeInTheDocument();
+  expect(await screen.findByText(`${dateString} – Pizza`)).toBeInTheDocument();
+});
+
