+++ src/pages/_app.tsx
@@ -1,6 +1,12 @@
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
+import { useEffect } from "react";
+import { auth } from "@/lib/firebaseClient";
+import { signInAnonymously } from "firebase/auth";
 
 export default function App({ Component, pageProps }: AppProps) {
+  useEffect(() => {
+    signInAnonymously(auth).catch(() => {});
+  }, []);
   return <Component {...pageProps} />;
 }
